+++ src/IO/FileAnalyse/SHPFileAnalyse.cpp
@@ -155,7 +155,7 @@ IO::FileAnalyse::FrameDetail *IO::FileAnalyse::SHPFileAnalyse::GetFrameDetail(UO
 	if (pack == 0)
 		return 0;
 	
-	NEW_CLASS(frame, IO::FileAnalyse::FrameDetail(pack->fileOfst, (UInt32)pack->packSize));
+	NEW_CLASS(frame, IO::FileAnalyse::FrameDetail(pack->fileOfst, pack->packSize));
 	sptr = Text::StrUOSInt(Text::StrConcatC(sbuff, UTF8STRC("Packet ")), index);
 	frame->AddHeader(CSTRP(sbuff, sptr));
