+++ src/IO/FileAnalyse/SMBIOSFileAnalyse.cpp
@@ -154,7 +154,7 @@ IO::FileAnalyse::FrameDetail *IO::FileAnalyse::SMBIOSFileAnalyse::GetFrameDetail
 	}
 
 
-	NEW_CLASS(frame, IO::FileAnalyse::FrameDetail(pack->fileOfst, (UInt32)pack->packSize));
+	NEW_CLASS(frame, IO::FileAnalyse::FrameDetail(pack->fileOfst, pack->packSize));
 	frame->AddUIntName(0, 1, CSTR("Type"), pack->packType, SMBIOSTypeGetName(pack->packType));
 	frame->AddUInt(1, 1, CSTR("Length"), packBuff[1]);
 	frame->AddHex16(2, CSTR("Handle"), ReadUInt16(&packBuff[2]));
